chore(Arena): cleanup the Arena.hpp and add Container.hpp so container can use the permanent and temporary arenas
