--- src/Arena/memory.cpp
@@ -1,29 +1,0 @@
-#include "memory.hpp"
-
-#include <cstddef>
-
-Arena::Arena(size_t size)
-  : memory_(::new unsigned char[size]), size_(size)
-{
-  fill_ = memory_;
-}
-
-Arena::~Arena() {
-  delete[] memory_;
-}
-
-void Arena::clear() {
-  fill_ = memory_;
-}
-
-ArenaSnapshot Arena::snapshot() {
-  return fill_;
-}
-
-void Arena::rollback(ArenaSnapshot snapshot) {
-  fill_ = snapshot;
-}
-
-uintptr_t Arena::align(uintptr_t ptr, size_t alignment) {
-  return (ptr + alignment - 1) & ~(alignment - 1);
-}
